requirements.txt
asgiref==3.6.0
dj-database-url==1.2.0
dj-static==0.0.6
Django==4.1.7
django-paginator==0.2a3
django-taggit==3.1.0
gunicorn==20.1.0
Pillow==9.4.0
sqlparse==0.4.3
static3==0.7.0
tzdata==2022.7
whitenoise==6.4.0
